test-ts: combine tests into DarknodePayment.ts
